chore(sdk-coin-sol): move account-lib tests
- moves tests from account-lib
- updates imports to get tests working again

Ticket: BG-48156
