+++ modules/sdk-coin-sol/test/resources/sol.ts
